+++ History.md
@@ -3,7 +3,6 @@
 ## v0.6.0
 
 - Code Format Refactor
-- Bumped minimum package API version to Meteor 1.3 (to support Tracker)
 - Changed Deps to Tracker (#49)
 - Only show log output if running in development
 - Added _timeSync Meteor Method for doing timesync over DDP instead of HTTP
@@ -12,7 +11,7 @@
 - Shortened resync interval from 1 minute to 30 seconds when using DDP.
 - Added tests for DDP and HTTP sync
 - Added option to set the timesync URL using `TimeSync.setSyncUrl`
-- ... more to come, v0.6.0 is a work in progress!
+- Removed IE8 compat function
 
 ## v0.5.4